Dedupe & fix tests from merge-forward
